+++ ui/ui-expenses/src/screens/expense/index.js
@@ -1,0 +1,1 @@
+export { default } from './expense.container'
