--- ui/ui-expenses/src/screens/prices/index.js
@@ -1,1 +1,0 @@
-export { default } from './prices.container'
