+++ ui/ui-expenses/src/store/drivers/router.js
@@ -3,8 +3,7 @@ import { router } from '@k-redux-router/react-k-ramel'
 const routes = {
   '/': {
     code: 'list',
-    '/expense': { code: 'create' },
-    '/expense/:id': { code: 'expense' },
+    '/expense': { code: 'expense' },
   },
 }
